Changed to storing SBERT and OPENAI embeddings as separate columns in the same row of the embedding table.
This denormalization simplifies the code considerably.  It's highly unlikely that we will get an explosion of embeddings.   More likely we will pick one, and stick with it.

Also fixed a bug caused because we assumed the langchain splits of a document were contiguous.  It turns out they are not in every case.